CTSDelighter turned into floatingDelighter module. CSS now piped through the module.
