--- CancerGov/_src/Scripts/NCI/UX/Common/Enhancements/clinicalTrialsDelighter.js
@@ -1,60 +1,0 @@
-define(function(require) {
-    var $ = require('jquery');
-	var AdobeAnalytics = require('Patches/AdobeAnalytics');
-	var ctsPath = '/about-cancer/treatment/clinical-trials/search';
-	var pathName = location.pathname.replace(/\/$/, "").toLowerCase();
-    
-    function _initialize() {    
-        /* Only display the delighter on the NCI Home page but not on other
-        * pages regardless if they are using the home page template
-        * ---------------------------------------------------------------- */
-        if( $("body").hasClass("ncihome") ) {
-            var delighter = $('<div id="delighter-homePage"><a href="' + ctsPath + '">Find a <br/>Clinical Trial</a></div>');
-            delighter.find('a').on('click.analytics',function(e){
-                var s = AdobeAnalytics.getSObject();
-                NCIAnalytics.HomePageDelighterClick($(this), 'hp_find', s.pageName);
-            });
-            delighter.insertAfter('#PageOptionsControl1');
-        }
-
-        /* Display the delighter on all pages under the /clinical-trials path, except for the search/results/view pages
-        * ---------------------------------------------------------------- */
-        else if( pathName.indexOf("about-cancer/treatment/clinical-trials") > -1 ) {
-            // if(pathName.indexOf(ctsPath) == -1 && pathName.indexOf('advanced-search') == -1 || /\/v$/.test(pathName))
-            if(pathName.indexOf(ctsPath) == -1 && pathName.indexOf('advanced-search') == -1 ||
-            pathName.match('/about-cancer/treatment/clinical-trials/search/trial-guide') ||
-            pathName.match('/about-cancer/treatment/clinical-trials/search/help'))
-            {
-                var delighter = $('<div id="delighter-innerpage"><a href="' + ctsPath + '">Find a <br/>Clinical Trial</a></div>');
-                delighter.find('a').on('click.analytics',function(e){
-                    var s = AdobeAnalytics.getSObject();
-                    NCIAnalytics.HomePageDelighterClick($(this), 'hp_find', s.pageName);
-                });
-                delighter.insertAfter('#PageOptionsControl1');
-            }
-        }
-    }
-
-	/**
-	 * Identifies if this enhancement has been initialized or not.
-	 * @type {Boolean}
-	 */
-	var initialized = false;
-
-	/**
-	 * Exposed functions available to this module.
-	 */
-	return {
-		init: function() {
-			if (initialized) {
-				return;
-			}
-
-			_initialize();
-
-			initialized = true;
-		}
-	};
-
-});
-
