+++ CancerGov/_src/Scripts/NCI/UX/PageSpecific/Landing/LandingPage.js
@@ -3,6 +3,6 @@ define(function(require) {
 	require('Modules/carousel/carousel');
 
 	$(function() {
-		require('Common/Enhancements/clinicalTrialsDelighter').init();
+		require('Modules/floatingDelighter').init();
 	});
 });
