+++ CancerGov/_src/StyleSheets/AppModuleSpecific/cts/_NCI.Delighters.scss
@@ -272,136 +272,4 @@
 }
 
 
-// the following code is for the purple box that links to the CTS section on home page and elsewhere. This keeps if from colliding with social medial links when minimizing screen.
-#delighter-homePage, #delighter-innerpage {
-  //display: none;
-  position:fixed;
-  top:51%;
-  right:0;
-  z-index: 998;
-
-  a {
-    display:inline-block;
-    color:#fff;
-    font:bold 14px/16px Montserrat;
-    background-color: #82368C;
-    padding: 7px 10px;
-    max-width: 115px;
-    text-align: center;
-
-    &:hover {
-      background-color: #3f0557;
-      text-decoration: none;
-    }
-  }
-}
-
-/* Large Screens */
-@include bp('(min-width:1025px)') {
-  #delighter-homePage, #delighter-innerpage {
-    // left:0;
-    // max-width: 1250px;
-    //width: 100%;
-    text-align: right;
-    margin:0 auto;
-    display: inline;
-    right: 1%;
-    width: 110px;
-  }
-}
-/* Extra Large Screens */
-@include bp('(min-width:1770px)') {
-  #delighter-homePage, #delighter-innerpage {
-    right: 10%;
-  }
-}
-/* Tablet/Mobile styles */
-@include bp(medium-down) {
-  #delighter-homePage, #delighter-innerpage {
-    position: relative;
-    float: right;
-    top: 13px;
-    right: 10px;
-  }
-  #delighter-innerpage {
-    margin-top: -65px;
-  }
-}
-
-/* Mobile styles */
-@include bp(small) {
-  #delighter-homePage, #delighter-innerpage {
-    right:5px;
-  }
-}
-
-// to fix find a clinical trial button positioning at small widths
-@include bp('(max-width: 360px)') {
-  #delighter-homePage, #delighter-innerpage {
-    top: 0;
-    margin-bottom: 8px;
-  }
-}
-
-// the following code is for the purple box that links to the CTS section on home page and elsewhere. This keeps if from colliding with social medial links when minimizing screen.
-@media only screen and (max-width: 400px) {
-  #delighter-innerpage {
-    float: none;
-    margin-top: 10px;
-    position: static;
-    margin-bottom: 8px;
-    margin-right: 5px;
-    text-align: right;
-  }
-}
-
-@include bp(xtra-small) {
-  #delighter-homePage {
-    position: initial;
-    display: block;
-    float: none;
-    z-index: 0;
-    width: 100%;
-    margin: 0;
-    margin-bottom: 10px;
-    padding: 0;
-    
-    a {
-      display: block;
-      width: 95%;
-      max-width: 100%;
-      margin: 0 auto;
-      padding: 15px 0;
-      font-size: 1.2em;
-      letter-spacing: 1.5px;
-      font-weight: normal;
-      box-shadow: 0 10px 4px -8px rgba(0,0,0,0.4);
-      transition: all 150ms ease-in-out;
-      
-      &:hover, &:focus, &:active {
-        text-decoration: none;
-      }
-      
-      &:hover {
-        box-shadow: 0 15px 6px -12px rgba(0,0,0,0.4);
-      }
-
-      &:active {
-        box-shadow: 0 10px 5px -10px rgba(0,0,0,0.4);
-      }
-
-      br {
-        display: none;
-      }
-    }
-  }
-}
-
-// 10/31/2017 hotfix to hide the feedback delighter upon launch of new survey tool
-.results-delighter-rail,
-.view-delighter-rail {
-  .delighter.cts-feedback {
-    display: none;
-  }
-}
 /****************** End Delighters styling ***************** */
